fix: p5_stubs sketch pokes plain I2S CONF1 regs; battery 61/0/0
chore: retire clippy tech debt, -D warnings green across workspace

Fold the triplicated deep-sleep tick into Esp32S3::sleep_tick;
drop identity ops (BASE+0x00, (0<<n), 0|x), same-type casts,
needless borrows/refs/parens/muts, format!(static), range-loop
indexing, dead consts/vars/imports; merge run_flash injection nests
via let-chains (kept the pin>=32 guard via let-chain). No behavior
change: 37/37 suites green, fmt/wasm32 clean, battery smoke green.

+++ crates/esp32s3-soc/src/sdmmc.rs
@@ -863,7 +863,7 @@ mod tests {
         issue(&mut d, 24, 0, true, true, true); // WRITE_BLOCK
         // Write a recognizable pattern.
         for i in 0..128u32 {
-            d.write32(FIFO, i as u32 * 0x0101_0101);
+            d.write32(FIFO, i * 0x0101_0101);
         }
         assert!(d.read32(RINTSTS) & INT_DATA_OVER != 0);
         // Now read it back.
@@ -872,7 +872,7 @@ mod tests {
         let mut mismatch = 0;
         for i in 0..128u32 {
             let w = d.read32(FIFO);
-            if w != i as u32 * 0x0101_0101 {
+            if w != i * 0x0101_0101 {
                 mismatch += 1;
             }
         }
